+++ src/packaging_assistant/modules/structure/__init__.py
@@ -1,6 +1,6 @@
 from .legacy import LegacyDielineAdapter
 from .models import StructureGeneration, StructureSpec
-from .registry import MODELS, model_report, resolve_model
+from .registry import MODELS, model_choices, model_report, resolve_model
 from .service import generate_structure_template, write_structure_outputs
 
 __all__ = [
@@ -9,6 +9,7 @@ __all__ = [
     "StructureGeneration",
     "StructureSpec",
     "generate_structure_template",
+    "model_choices",
     "model_report",
     "resolve_model",
     "write_structure_outputs",